refactor: split timeline components
