+++ libs/application/templates/parental-leave/src/fields/components/Timeline/Timeline.tsx
@@ -1,0 +1,67 @@
+import React, { FC } from 'react'
+import { useWindowSize } from 'react-use'
+
+import { Box } from '@island.is/island-ui/core'
+import { theme } from '@island.is/island-ui/theme'
+
+import { Panel } from './Panel'
+import { Chart } from './Chart'
+import * as styles from './Timeline.treat'
+
+export interface TimelinePeriod {
+  actualDob?: boolean
+  startDate: string
+  endDate: string
+  ratio: string
+  duration: string
+  title: string
+  color?: string
+  canDelete?: boolean
+}
+
+interface TimelineProps {
+  editable?: boolean
+  initDate: Date
+  periods: TimelinePeriod[]
+  spanInMonths?: number
+  title: string
+  titleSmall?: string
+  onDeletePeriod?: (index: number) => void
+}
+
+export const Timeline: FC<TimelineProps> = ({
+  editable = true,
+  initDate,
+  periods,
+  spanInMonths = 18,
+  title,
+  titleSmall,
+  onDeletePeriod,
+}) => {
+  const { width } = useWindowSize()
+  const isMobile = width < theme.breakpoints.md
+  const dayWidth = isMobile ? 2 : 3
+
+  return (
+    <Box display="flex" width="full" position="relative">
+      <Panel
+        initDate={initDate}
+        title={title}
+        titleSmall={titleSmall || title}
+        periods={periods}
+        isMobile={isMobile}
+        editable={editable}
+        onDeletePeriod={onDeletePeriod}
+      />
+
+      <Chart
+        initDate={initDate}
+        periods={periods}
+        dayWidth={dayWidth}
+        spanInMonths={spanInMonths}
+      />
+
+      <Box className={styles.scrollGradient} />
+    </Box>
+  )
+}
